fix: admin controller, service test 모두 성공 되도록 수정.

+++ src/main/java/com/example/gamecommunity/domain/admin/service/AdminService.java
@@ -10,6 +10,7 @@ import com.example.gamecommunity.domain.post.dto.PostResponseDto;
 import com.example.gamecommunity.domain.post.entity.Post;
 import com.example.gamecommunity.domain.post.repository.PostRepository;
 import com.example.gamecommunity.domain.user.entity.User;
+import com.example.gamecommunity.domain.user.entity.UserRoleEnum;
 import com.example.gamecommunity.domain.user.repository.UserRepository;
 import com.example.gamecommunity.global.exception.common.BusinessException;
 import com.example.gamecommunity.global.exception.common.ErrorCode;
@@ -28,11 +29,21 @@ public class AdminService {
   private final UserRepository userRepository;
   private final PostRepository postRepository;
 
-  public List<AdminUserResponseDto> getUsers() {
+  public List<AdminUserResponseDto> getUsers(UserDetailsImpl userDetails) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     return userRepository.findAll().stream().map(AdminUserResponseDto::new).toList();
   }
 
-  public AdminUserResponseDto getUser(long userId) {
+  public AdminUserResponseDto getUser(UserDetailsImpl userDetails, long userId) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     var user = userRepository.findById(userId).orElseThrow(
         () -> new BusinessException(HttpStatus.NOT_FOUND, ErrorCode.NOT_FOUND_USER_EXCEPTION)
     );
@@ -40,16 +51,26 @@ public class AdminService {
     return new AdminUserResponseDto(user);
   }
 
-  public void deleteUser(long userId) {
+  public void deleteUser(UserDetailsImpl userDetails, long userId) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     userRepository.findById(userId).orElseThrow(
-        ()-> new BusinessException(HttpStatus.NOT_FOUND, ErrorCode.NOT_FOUND_USER_EXCEPTION)
+        () -> new BusinessException(HttpStatus.NOT_FOUND, ErrorCode.NOT_FOUND_USER_EXCEPTION)
     );
 
     userRepository.deleteById(userId);
   }
 
   @Transactional
-  public void setBlock(UserBlockRequestDto userBlockRequestDto) {
+  public void setBlock(UserDetailsImpl userDetails, UserBlockRequestDto userBlockRequestDto) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     var userId = userBlockRequestDto.userId();
     var blockDate = userBlockRequestDto.blockDate();
 
@@ -60,13 +81,23 @@ public class AdminService {
     user.setBlockDate(blockDate.toInstant(ZoneOffset.UTC));
   }
 
-  public List<PostResponseDto> getReportedPosts() {
+  public List<PostResponseDto> getReportedPosts(UserDetailsImpl userDetails) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     return postRepository.findAll().stream()
         .filter(p -> p.getReport() > 0)
         .map(PostResponseDto::fromEntity).toList();
   }
 
-  public PostResponseDto getReportedPost(long postId) {
+  public PostResponseDto getReportedPost(UserDetailsImpl userDetails, long postId) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     var post = postRepository.findById(postId).orElseThrow(
         () -> new BusinessException(HttpStatus.NOT_FOUND, ErrorCode.NOT_FOUND_POST_EXCEPTION)
     );
@@ -76,13 +107,18 @@ public class AdminService {
 
   @Transactional
   public void writeNotice(
+      UserDetailsImpl userDetails,
       PostRequestDto requestDto,
       GameType gameType,
-      GameName gameName,
-      User user
+      GameName gameName
   ) {
+    if (userDetails.getUser().getRole() != UserRoleEnum.ADMIN) {
+      throw new BusinessException(HttpStatus.BAD_REQUEST,
+          ErrorCode.AUTHENTICATION_MISMATCH_EXCEPTION);
+    }
+
     Post post = new Post(requestDto, gameType, gameName, BoardName.NOTICE_BOARD, "",
-        user);
+        userDetails.getUser());
     postRepository.save(post);
   }
 }
